:bug: Fix segfault when using MinSizeRel
MinSizeRel will use fpu registers in initialize_platform which causes
a segfault because the FPU has yet to be initialized. By using a new
smaller function with just calls to init data section and FPU, the
FPU registers are not used. This solution seems to work for now.

Rename initialize_target to initialize_platform.

Make at_benchmark readable and understandable via the console.