:zap: Improve at::server_write() failure rate
In some situations, during a server write, a packet will be returned
before a "SEND OK" response has been emitted. This is known and expected
behavior of the AT command system with ESP devices. The problem with
this behavior is that, the previous write function would toss out the
packet contents, reading until it reached the "SEND OK" message. This
will tend to cause the server_read() to read back nothing forever as the
previous packet has already been tossed out.

at::server_write() now searches for both the "SEND OK" response, as well
as the the start of a packet. If it finds either one, it returns a
successful function call. The packet manager has its state updated to
expect the section of a packet right after the start ensuring that
subsequent read operations get the packet stored in the internal working
buffer. This has helped to improve the number of read timeouts.

Bump version to 2.0.0-alpha.2